fix: 여행방에 이미 초대된 사용자 검색 제외하도록 수정
