+++ src/main/java/com/example/travelday/domain/auth/service/MemberManageService.java
@@ -45,15 +45,4 @@ public class MemberManageService {
 
         member.updateNickname(nickname);
     }
-
-    @Transactional
-    public List<MemberInfoResDto> searchMembers(String keyword) {
-        Pageable pageable = PageRequest.of(0, 5);
-
-        Page<Member> members = memberRepository.findByNicknameContaining(keyword, pageable);
-
-        return members.stream()
-                .map(MemberInfoResDto::of)
-                .toList();
-    }
 }
