+++ src/main/java/com/example/travelday/domain/travelroom/controller/TravelRoomController.java
@@ -1,5 +1,6 @@
 package com.example.travelday.domain.travelroom.controller;
 
+import com.example.travelday.domain.auth.dto.response.MemberInfoResDto;
 import com.example.travelday.domain.travelroom.dto.request.TravelRoomReqDto;
 import com.example.travelday.domain.travelroom.dto.response.TravelRoomMembersResDto;
 import com.example.travelday.domain.travelroom.dto.response.TravelRoomResDto;
@@ -66,5 +67,13 @@ public class TravelRoomController {
     public ResponseEntity<ApiResponseEntity<String>> updateTravelRoom(@PathVariable Long travelRoomId, @RequestBody TravelRoomReqDto requestDto, @AuthenticationPrincipal UserDetails userDetails) {
         travelRoomService.updateTravelRoom(travelRoomId, requestDto, userDetails.getUsername());
         return ResponseEntity.ok(ApiResponseEntity.of(ResponseText.SUCCESS_UPDATE_TRAVELROOM));
+    }
+
+    /**
+     * 초대할 수 있는 사용자 검색
+     */
+    @GetMapping("/{travelRoomId}/user/search")
+    public ResponseEntity<ApiResponseEntity<List<MemberInfoResDto>>> searchMember(@RequestParam String keyword) {
+        return ResponseEntity.ok(ApiResponseEntity.of(travelRoomService.searchMembers(keyword)));
     }
 }
